refactor: move session serve to a separete file
just preparing for the actual changes
